+++ CHANGELOG.md
@@ -6,9 +6,16 @@
 
 ##### Enhancements
 
-* Add search function to `apple` theme.
+* Add search function to `apple` theme.  
   [Giles Payne](https://github.com/komakai)
   [#726](https://github.com/realm/jazzy/issues/726)
+
+* Add option `--[no-]separate-global-declarations` to always create separate
+  documentation pages for top-level declarations as well as classes,
+  structures, enums etc. even if they don't have members. As part of this,
+  improve the main page declaration in all modes.  
+  [Nikolay Volosatov](https://github.com/bamx23)
+  [John Fairhurst](https://github.com/johnfairh)
 
 ##### Bug Fixes